// frontend/src/api/model/githubCom4H1RZooraInternalDomainInvoiceStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainInvoiceStatus = typeof GithubCom4H1RZooraInternalDomainInvoiceStatus[keyof typeof GithubCom4H1RZooraInternalDomainInvoiceStatus];


export const GithubCom4H1RZooraInternalDomainInvoiceStatus = {
  InvoiceStatusDraft: 'draft',
  InvoiceStatusPending: 'pending',
  InvoiceStatusPaid: 'paid',
  InvoiceStatusCanceled: 'canceled',
  InvoiceStatusExpired: 'expired',
  InvoiceStatusRefunded: 'refunded',
} as const;